+++ CHANGELOG.md
@@ -6,6 +6,7 @@ All notable changes to `dropblockeditor` will be documented in this file.
 
 - upgrade to Livewire v3
 - `php artisan dropblockeditor:make` puts Livewire component in `App\Livewire` namespace unless otherwise configured
+- `php artisan dropblockeditor:make` now assumes you want an edit component by default
 
 ## 0.1.3 - 2022-06-23